moved global functions to helpers.py
